Allow tmp root to vary per system
mktemp uses $TMPDIR as the tmp root if it is set and defaults to /tmp
if it is not. So there is no reason to specify --tmpdir=/tmp. On Linux
TMPDIR is not set by default, and /tmp will be used. On OSX it is set
and uses some user-specific location. On TeamCity $TMPDIR is set to a
build-specific location.

It's actually the TeamCity behaviour that we're doing this for. Using
/tmp there is problematic because that is very small on our build
agents, so we need to use $TMPDIR instead.